backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pagehound"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    "ebooklib>=0.18",
    "pymupdf>=1.24",
    "python-multipart>=0.0.18",
    "aiofiles>=24.1",
    "Pillow>=11.0",
    "lxml>=5.3",
    "isbnlib>=3.10",
    "websockets>=14.1",
    "greenlet>=3.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "httpx>=0.28",
    "anyio>=4.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
